+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/replication/ReplicatorBuilder.java
@@ -33,6 +33,7 @@ public abstract class ReplicatorBuilder<S, T, E> {
 
     private T target;
     private S source;
+    private int id = BasicReplicator.NULL_ID;
     private List<HttpConnectionRequestInterceptor> requestInterceptors = new ArrayList
             <HttpConnectionRequestInterceptor>();
     private List<HttpConnectionResponseInterceptor> responseInterceptors = new ArrayList
@@ -55,8 +56,7 @@ public abstract class ReplicatorBuilder<S, T, E> {
             pushReplication.target = super.target;
             pushReplication.responseInterceptors.addAll(super.responseInterceptors);
             pushReplication.requestInterceptors.addAll(super.requestInterceptors);
-            return new BasicReplicator(pushReplication);
-
+            return new BasicReplicator(pushReplication, super.id);
         }
     }
 
@@ -91,7 +91,7 @@ public abstract class ReplicatorBuilder<S, T, E> {
                 pullReplication.filter = filter;
             }
 
-            return new BasicReplicator(pullReplication);
+            return new BasicReplicator(pullReplication, super.id);
         }
 
         /**
@@ -127,6 +127,12 @@ public abstract class ReplicatorBuilder<S, T, E> {
      */
     public E from(S source) {
         this.source = source;
+        //noinspection unchecked
+        return (E) this;
+    }
+
+    public E withId(int id) {
+        this.id = id;
         //noinspection unchecked
         return (E) this;
     }